+++ CHANGELOG.md
@@ -27,3 +27,5 @@
 ## 0.0.8
 - Calculate two-hour period lucky list.
 - Calculate two-hour period list.
+
+## 0.0.9